refactor: count 반환 값을 Long에서 int로 수정

+++ bucketback-api/src/main/java/com/programmers/bucketback/domains/review/api/dto/response/ReviewGetByCursorResponse.java
@@ -8,7 +8,7 @@ import com.programmers.bucketback.domains.review.model.ReviewCursorSummary;
 
 public record ReviewGetByCursorResponse(
 	String nextCursorId,
-	Long itemReviewTotalCount,
+	int itemReviewTotalCount,
 	int totalCount,
 	List<ReviewCursorSummary> reviews
 ) {
