+++ bucketback-domain/src/main/java/com/programmers/bucketback/domains/review/repository/ReviewRepositoryForCursor.java
@@ -12,5 +12,5 @@ public interface ReviewRepositoryForCursor {
 		int pageSize
 	);
 
-	Long getReviewCount(Long itemId);
+	int getReviewCount(Long itemId);
 }
